+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Changelog
 -----
 
 - Workaround DNSSEC testing bug in Debian for some domains
-  - On Windows, querying for a `DNSKEY` record on `proton.ch` will return a `RRSET` and `RRSIG`. However, running the same query on
+  - On Windows, querying for a `DNSKEY` record on `proton.ch` will return a `RRSET` and `RRSIG`. However, running the same query on 
     Linux will only return a `RRSET`, but will return a `RRSET` and `RRSIG` if another record type is requested, such
     as `A`